remove runCatching, remove bloat in parsing of response
