+++ CorePayments/src/main/java/com/paypal/android/corepayments/model/AppSwitchEligibility.kt
@@ -1,0 +1,10 @@
+package com.paypal.android.corepayments.model
+
+import androidx.annotation.RestrictTo
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+data class AppSwitchEligibility(
+    val appSwitchEligible: Boolean,
+    val launchUrl: String?,
+    val ineligibleReason: String?
+)
